zjugit-script/pyproject.toml
[project]
name = "zjugit-script"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-gitlab>=6.3.0",
    "pyyaml>=6.0.2",
]
